utils: trace: Fix traces number increasing in waitForCondition().
Before this commit, the total number of traces was computed from Added event.
This reveals to be a bad idea as it was possible to return earlier from
waitForCondition() because the "total" number of traces would equal the watched
number of traces.
So, this commit fixes this behavior by comparing the watched number of traces
with the number of traces which have the given ID.

Fixes: a6bb14e8ff0f ("utils: trace: Add new API to interact with trace.")
Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>